refactor(types): centralize shared frontend types under src/types with a barrel
