+++ frontend/src/components/CourseCardOverlay.vue
@@ -154,7 +154,7 @@ import type {
 	CourseInstructorInfo,
 	Resource,
 	SessionUser,
-} from '@/types/api'
+} from '@/types'
 
 const router = useRouter()
 const user = inject<SessionUser>('$user')!
