+++ frontend/src/components/Settings/Badges/BadgeForm.vue
@@ -88,7 +88,7 @@ import { Combobox, Button, call, FormControl, toast } from 'frappe-ui'
 import BooleanSwitch from '@/components/Controls/BooleanSwitch.vue'
 import { computed, ref, watch } from 'vue'
 import { cleanError } from '@/utils'
-import type { Badges, Badge } from '@/components/Settings/types'
+import type { Badges, Badge } from '@/types'
 import CodeEditor from '@/components/Controls/CodeEditor.vue'
 import Uploader from '@/components/Controls/Uploader.vue'
 import SettingsLayout from '@/components/Layouts/SettingsLayout.vue'
