+++ frontend/src/components/Settings/Coupons/CouponDetails.vue
@@ -87,7 +87,7 @@
 import { Button, FormControl, toast, createDocumentResource } from 'frappe-ui'
 import BooleanSwitch from '@/components/Controls/BooleanSwitch.vue'
 import { computed, reactive } from 'vue'
-import type { Coupon, Coupons } from './types'
+import type { Coupon, Coupons } from '@/types'
 import CouponItems from '@/components/Settings/Coupons/CouponItems.vue'
 import SettingsLayout from '@/components/Layouts/SettingsLayout.vue'
 import Select from '@/components/Controls/Select.vue'
@@ -101,9 +101,8 @@ const props = defineProps<{
 
 const isNew = !props.data?.name
 
-// New coupons are edited on a local object; existing ones are loaded as a full
-// document (child table + `modified`) so a single save persists everything and
-// the document resource keeps `modified` in sync across saves.
+// New coupons edit a local object; existing ones load as a full document (child
+// table + `modified`) so one save persists all of it and `modified` stays in sync.
 const localDoc = reactive<any>({
 	enabled: true,
 	discount_type: 'Percentage',
