+++ frontend/src/components/Settings/EmailTemplate/EmailTemplateAdd.vue
@@ -75,7 +75,7 @@ import {
 } from 'frappe-ui'
 import { computed, reactive, ref } from 'vue'
 import { cleanError } from '@/utils'
-import type { EmailTemplate, EmailTemplateStep } from '@/types/email'
+import type { EmailTemplate, EmailTemplateStep } from '@/types'
 import RichTextEditor from '@/components/RichTextEditor.vue'
 
 interface P {
