+++ frontend/src/components/Settings/Members.vue
@@ -137,7 +137,7 @@ import { List, ListCell, ListRow, ListRows } from 'frappe-ui/list'
 import { useRouter } from 'vue-router'
 import { ref, computed, watch, inject } from 'vue'
 import { useOnboarding, useTelemetry } from 'frappe-ui/frappe'
-import type { User } from '@/components/Settings/types'
+import type { User } from '@/types'
 import NewMemberModal from '@/components/Modals/NewMemberModal.vue'
 import SettingsLayout from '@/components/Layouts/SettingsLayout.vue'
 import EmptyStateLayout from '@/components/Layouts/EmptyStateLayout.vue'
