+++ frontend/src/pages/Programs/ProgramProgressSummary.vue
@@ -74,7 +74,7 @@ import {
 	ListView,
 	NumberChart,
 } from 'frappe-ui'
-import type { ProgramMember } from './types'
+import type { ProgramMember } from '@/types'
 import { computed, ref, watch } from 'vue'
 
 const show = defineModel<boolean>({ default: false })
