+++ frontend/src/types/coupons.ts
