+++ frontend/src/types/index.ts
@@ -1,0 +1,12 @@
+// Barrel for the hand-written type modules, so callers import from '@/types'.
+// The generated `lms/` doctype types stay on their own path — they are regenerated
+// wholesale and their names are not curated for collisions.
+export * from './api'
+export * from './coupons'
+export * from './email'
+export * from './notes'
+export * from './programmingExercises'
+export * from './programs'
+export * from './raven'
+export * from './settings'
+export * from './sidebar'
