+++ frontend/src/types/notes.ts
