+++ frontend/src/types/programmingExercises.ts
