+++ frontend/src/types/settings.ts
